Разбиение сервера на партиалы
